Habit completion state with strike-trough text and subtle opacity
