+++ lib/pages/habits/habits_page.dart
@@ -4,8 +4,8 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/blocs/habits/habits_cubit.dart';
 import 'package:lotti/blocs/habits/habits_state.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/widgets/charts/habits/dashboard_habits_chart.dart';
 import 'package:lotti/widgets/charts/utils.dart';
+import 'package:lotti/widgets/habits/habit_completion_card.dart';
 import 'package:lotti/widgets/habits/habit_page_app_bar.dart';
 import 'package:lotti/widgets/habits/habit_streaks.dart';
 import 'package:lotti/widgets/misc/timespan_segmented_control.dart';
@@ -53,24 +53,7 @@ class HabitsTabPage extends StatelessWidget {
                     ),
                   const SizedBox(height: 15),
                   ...state.openNow.map((habitDefinition) {
-                    return HabitChartLine(
-                      habitDefinition: habitDefinition,
-                      rangeStart: rangeStart,
-                      rangeEnd: rangeEnd,
-                      showGaps: showGaps,
-                    );
-                  }),
-                  if (state.completed.isNotEmpty)
-                    Padding(
-                      padding: const EdgeInsets.only(top: 20),
-                      child: Text(
-                        localizations.habitsCompletedHeader,
-                        style: chartTitleStyle(),
-                      ),
-                    ),
-                  const SizedBox(height: 15),
-                  ...state.completed.map((habitDefinition) {
-                    return HabitChartLine(
+                    return HabitCompletionCard(
                       habitDefinition: habitDefinition,
                       rangeStart: rangeStart,
                       rangeEnd: rangeEnd,
@@ -87,7 +70,24 @@ class HabitsTabPage extends StatelessWidget {
                     ),
                   const SizedBox(height: 15),
                   ...state.pendingLater.map((habitDefinition) {
-                    return HabitChartLine(
+                    return HabitCompletionCard(
+                      habitDefinition: habitDefinition,
+                      rangeStart: rangeStart,
+                      rangeEnd: rangeEnd,
+                      showGaps: showGaps,
+                    );
+                  }),
+                  if (state.completed.isNotEmpty)
+                    Padding(
+                      padding: const EdgeInsets.only(top: 20),
+                      child: Text(
+                        localizations.habitsCompletedHeader,
+                        style: chartTitleStyle(),
+                      ),
+                    ),
+                  const SizedBox(height: 15),
+                  ...state.completed.map((habitDefinition) {
+                    return HabitCompletionCard(
                       habitDefinition: habitDefinition,
                       rangeStart: rangeStart,
                       rangeEnd: rangeEnd,
